Add proceedural selection sort
I tried to introspect the function name inside the generic delegate
object typology I am using but have not had success.  I cannot get the
generic to yield an object that is similar enough to the delegate to
simply access `alg.[...].Method.Name`.  The closest I could get was by
wrapping the delegate in an `Expression` object, but beside being obtuse
and inflexible this did not completely work with the generic anyway.

For now I am simply passing a title parameter along with the algorithm
function itself to the algorithm executor.